Separate concerns between metrics views and tasks
